// src/types/personDetail.ts
export interface PersonDetail {
    birthday: string;
    known_for_department: string;
    deathday?: any;
    id: number;
    name: string;
    also_known_as: string[];
    gender: number;
    biography: string;
    popularity: number;
    place_of_birth: string;
    profile_path: string;
    adult: boolean;
    imdb_id: string;
    homepage?: any;
}